Create a module: common and move common things there
