+++ app/build.gradle
@@ -48,6 +48,7 @@ android {
 }
 
 dependencies {
+    implementation project(":core:common")
     implementation "androidx.media:media:$media_api_version"
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:1.8.0"
     implementation "androidx.appcompat:appcompat:$appcompat_version"
