+++ app/src/main/java/com/yurii/youtubemusic/ShareContentModalSheet.kt
@@ -15,8 +15,8 @@ import androidx.fragment.app.FragmentManager
 import coil.load
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import com.google.android.material.button.MaterialButton
+import com.youtubemusic.core.common.toPx
 import com.yurii.youtubemusic.models.MediaItem
-import com.yurii.youtubemusic.ui.toPx
 import de.hdodenhof.circleimageview.CircleImageView
 import net.glxn.qrgen.android.QRCode
 import timber.log.Timber
