+++ app/src/main/java/com/yurii/youtubemusic/screens/youtube/VideoItemsListAdapter.kt
@@ -10,11 +10,11 @@ import androidx.paging.PagingDataAdapter
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
+import com.youtubemusic.core.common.getValueAnimator
 import com.yurii.youtubemusic.databinding.ItemVideoBinding
 import com.yurii.youtubemusic.models.VideoItem
 import com.yurii.youtubemusic.services.downloader.DownloadManager
 import com.yurii.youtubemusic.ui.DownloadButton
-import com.yurii.youtubemusic.ui.getValueAnimator
 
 class VideoItemsListAdapter(private val callback: Callback) :
     PagingDataAdapter<VideoItem, VideoItemsListAdapter.MyViewHolder>(Comparator) {
