+++ app/src/main/java/com/yurii/youtubemusic/services/media/MediaServiceConnection.kt
@@ -8,11 +8,11 @@ import android.support.v4.media.MediaBrowserCompat
 import android.support.v4.media.MediaMetadataCompat
 import android.support.v4.media.session.MediaControllerCompat
 import android.support.v4.media.session.PlaybackStateCompat
+import com.youtubemusic.core.common.parcelable
+import com.youtubemusic.core.common.serializable
 import com.yurii.youtubemusic.models.*
 import com.yurii.youtubemusic.source.PlaylistRepository
 import com.yurii.youtubemusic.source.QueueModesRepository
-import com.yurii.youtubemusic.utilities.parcelable
-import com.yurii.youtubemusic.utilities.serializable
 import dagger.hilt.android.qualifiers.ApplicationContext
 import kotlinx.coroutines.flow.*
 import timber.log.Timber
