+++ app/src/main/res/layout/item_music.xml
@@ -6,7 +6,7 @@
 
     <data>
 
-        <import type="com.yurii.youtubemusic.utilities.UtilsKt" />
+        <import type="com.youtubemusic.core.common.FormatingUtilsKt" />
 
         <variable
             name="musicItem"
@@ -111,7 +111,7 @@
                 android:layout_marginStart="12dp"
                 android:layout_marginEnd="12dp"
                 android:layout_toStartOf="@id/more_options"
-                android:text="@{UtilsKt.parseDurationToHumanView(musicItem.durationInMillis)}"
+                android:text="@{FormatingUtilsKt.parseDurationToHumanView(musicItem.durationInMillis)}"
                 android:textSize="12sp"
                 tools:ignore="SmallSp"
                 tools:text="3:23" />
