+++ settings.gradle
@@ -16,3 +16,4 @@ dependencyResolutionManagement {
 
 include ':app'
 rootProject.name='YouTubeMusic'
+include ':core:common'
